+++ rumqttc/src/tls.rs
@@ -1,21 +1,39 @@
 use tokio::net::TcpStream;
+
+#[cfg(feature = "use-rustls")]
 use tokio_rustls::rustls;
+#[cfg(feature = "use-rustls")]
 use tokio_rustls::rustls::client::InvalidDnsNameError;
+#[cfg(feature = "use-rustls")]
 use tokio_rustls::rustls::{
     Certificate, ClientConfig, OwnedTrustAnchor, PrivateKey, RootCertStore, ServerName,
 };
+#[cfg(feature = "use-rustls")]
 use tokio_rustls::webpki;
-use tokio_rustls::{client::TlsStream, TlsConnector};
+#[cfg(feature = "use-rustls")]
+use tokio_rustls::TlsConnector as RustlsConnector;
 
-use crate::{Key, TlsConfiguration};
-
+#[cfg(feature = "use-rustls")]
+use crate::Key;
+#[cfg(feature = "use-rustls")]
 use std::convert::TryFrom;
-use std::io;
+#[cfg(feature = "use-rustls")]
 use std::io::{BufReader, Cursor};
-use std::net::AddrParseError;
+#[cfg(feature = "use-rustls")]
 use std::sync::Arc;
 
-/// TLS backend error
+use crate::framed::N;
+use crate::TlsConfiguration;
+
+#[cfg(feature = "use-native-tls")]
+use tokio_native_tls::TlsConnector as NativeTlsConnector;
+
+#[cfg(feature = "use-native-tls")]
+use tokio_native_tls::native_tls::{Error as NativeTlsError, Identity};
+
+use std::io;
+use std::net::AddrParseError;
+
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
     /// Error parsing IP address
@@ -24,28 +42,36 @@ pub enum Error {
     /// I/O related error
     #[error("I/O: {0}")]
     Io(#[from] io::Error),
+    #[cfg(feature = "use-rustls")]
     /// Certificate/Name validation error
     #[error("Web Pki: {0}")]
     WebPki(#[from] webpki::Error),
+    #[cfg(feature = "use-rustls")]
     /// Invalid DNS name
     #[error("DNS name")]
     DNSName(#[from] InvalidDnsNameError),
+    #[cfg(feature = "use-rustls")]
     /// Error from rustls module
     #[error("TLS error: {0}")]
     TLS(#[from] rustls::Error),
+    #[cfg(feature = "use-rustls")]
     /// No valid certificate in chain
     #[error("No valid certificate in chain")]
     NoValidCertInChain,
+    #[cfg(feature = "use-native-tls")]
+    #[error("Native TLS error {0}")]
+    NativeTls(#[from] NativeTlsError),
 }
 
-// The cert handling functions return unit right now, this is a shortcut
-impl From<()> for Error {
-    fn from(_: ()) -> Self {
-        Error::NoValidCertInChain
-    }
-}
+// // The cert handling functions return unit right now, this is a shortcut
+// impl From<()> for Error {
+//     fn from(_: ()) -> Self {
+//         Error::NoValidCertInChain
+//     }
+// }
 
-pub async fn tls_connector(tls_config: &TlsConfiguration) -> Result<TlsConnector, Error> {
+#[cfg(feature = "use-rustls")]
+pub async fn rustls_connector(tls_config: &TlsConfiguration) -> Result<RustlsConnector, Error> {
     let config = match tls_config {
         TlsConfiguration::Simple {
             ca,
@@ -118,19 +144,55 @@ pub async fn tls_connector(tls_config: &TlsConfiguration) -> Result<TlsConnector
             Arc::new(config)
         }
         TlsConfiguration::Rustls(tls_client_config) => tls_client_config.clone(),
+        #[allow(unreachable_patterns)]
+        _ => unreachable!("This cannot be called for other TLS backends than Rustls"),
     };
 
-    Ok(TlsConnector::from(config))
+    Ok(RustlsConnector::from(config))
+}
+
+#[cfg(feature = "use-native-tls")]
+pub async fn native_tls_connector(
+    tls_config: &TlsConfiguration,
+) -> Result<NativeTlsConnector, Error> {
+    let connector = match tls_config {
+        TlsConfiguration::SimpleNative { ca, der, password } => {
+            let cert = native_tls::Certificate::from_pem(ca)?;
+            let identity = Identity::from_pkcs12(der, password)?;
+            native_tls::TlsConnector::builder()
+                .add_root_certificate(cert)
+                .identity(identity)
+                .build()?
+        }
+        TlsConfiguration::Native => native_tls::TlsConnector::new()?,
+        #[allow(unreachable_patterns)]
+        _ => unreachable!("This cannot be called for other TLS backends than Native TLS"),
+    };
+
+    Ok(connector.into())
 }
 
 pub async fn tls_connect(
     addr: &str,
     port: u16,
     tls_config: &TlsConfiguration,
-) -> Result<TlsStream<TcpStream>, Error> {
-    let connector = tls_connector(tls_config).await?;
-    let domain = ServerName::try_from(addr)?;
+) -> Result<Box<dyn N>, Error> {
     let tcp = TcpStream::connect((addr, port)).await?;
-    let tls = connector.connect(domain, tcp).await?;
+
+    let tls: Box<dyn N> = match tls_config {
+        #[cfg(feature = "use-rustls")]
+        TlsConfiguration::Simple { .. } | TlsConfiguration::Rustls(_) => {
+            let connector = rustls_connector(tls_config).await?;
+            let domain = ServerName::try_from(addr)?;
+            Box::new(connector.connect(domain, tcp).await?)
+        }
+        #[cfg(feature = "use-native-tls")]
+        TlsConfiguration::Native | TlsConfiguration::SimpleNative { .. } => {
+            let connector = native_tls_connector(tls_config).await?;
+            Box::new(connector.connect(addr, tcp).await?)
+        }
+        #[allow(unreachable_patterns)]
+        _ => panic!("Unknown or not enabled TLS backend configuration"),
+    };
     Ok(tls)
 }